+++ library/src/main/java/com/twilio/video/TestUtils.java
@@ -2,4 +2,7 @@ package com.twilio.video;
 
 public class TestUtils {
     public static final long ICE_TIMEOUT = 20000;
+    public static final long SIP_TIMEOUT = 125;
+    public static final long STATE_TRANSITION_TIMEOUT = 20;
+    public static final long SMALL_WAIT = 5000;
 }