+++ CHANGELOG.md
@@ -1,4 +1,22 @@
 The Twilio Programmable Video SDKs use [Semantic Versioning](http://www.semver.org).
+###4.0.0
+
+Improvements
+
+- Added new state `Reconnecting` to `Room.State` and new callbacks `onReconnecting`, `onReconnected` to `Room.Listener`. When the `LocalParticipant` experiences a network interruption in signaling or media, the room will transition to `Reconnecting` and `Room.Listener` will notify the developer of this new state via `onReconnecting`. If the connection is successfully restored, `Room.Listener` will notify the developer via `onReconnected`. However, if the connection could not be reestablished `Room.Listener` will notify the developer via `onDisconnected`.
+
+Known issues
+
+- Network handoff, and subsequent connection renegotiation is not supported for IPv6 networks [#72](https://github.com/twilio/video-quickstart-android/issues/72)
+- Participant disconnect event can take up to 120 seconds to occur [#80](https://github.com/twilio/video-quickstart-android/issues/80) [#73](https://github.com/twilio/video-quickstart-android/issues/73)
+- The SDK is not side-by-side compatible with other WebRTC based libraries [#340](https://github.com/twilio/video-quickstart-android/issues/340)
+- Codec preferences do not function correctly in a hybrid codec Group Room with the following
+codecs:
+    - ISAC
+    - PCMA
+    - G722
+    - VP9
+
 ###3.2.2
 
 Improvements